add confirm delete and use db to schema

+++ public/js/delete.js
@@ -1,16 +1,20 @@
 const deleteHandler = async (event) => {
     event.preventDefault();
-    const id = window.location.toString().split('/')[
-        window.location.toString().split('/').length - 1
-      ];
-    const response = await fetch(`/api/avatar/${id}`, {
-        method: 'DELETE',
-        headers: {
-            'Content-type': 'application/json'
+    if (confirm("Are you sure you would like to delete this character?")) {
+        const id = window.location.toString().split('/')[
+            window.location.toString().split('/').length - 1
+          ];
+        const response = await fetch(`/api/avatar/${id}`, {
+            method: 'DELETE',
+            headers: {
+                'Content-type': 'application/json'
+            }
+        })
+        if (response.ok) {
+            document.location.replace(`/dashboard`)
         }
-    })
-    if (response.ok) {
-        document.location.replace(`/dashboard`)
+    } else {
+        return
     }
     
 }
